Move interfaces to model directory
