+++ back-end/src/model/github-api/UserMailScope.ts
